elf: fix typo in initial section offsets

+++ src/link/Elf/Object.zig
@@ -251,7 +251,7 @@ fn initOutputSection(self: Object, elf_file: *Elf, shdr: ElfShdr) error{OutOfMem
         .type = @"type",
         .flags = flags,
         .name = name,
-        .offset = std.math.maxInt(u32),
+        .offset = std.math.maxInt(u64),
     });
     return out_shndx;
 }
